requirements.txt
# Flask web framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3

# Database (SQLite is built-in, no extra package needed)
SQLAlchemy>=2.0.30

# Build tools
cmake
setuptools<81

# Face recognition
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.26.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
Werkzeug>=3.0.1
